Refactor stats to distinct mod and separate impls
